Rename project type
