+++ node-scripts/src/pipeline3/impl/dockerComposeFromArtifact.js
